spin out the capture actions in to a separate file
